+++ src/app/inventory/store/d2-store-factory.ts
@@ -1,11 +1,17 @@
 import { t } from 'app/i18next-t';
 import { armorStats } from 'app/search/d2-known-values';
-import { DestinyCharacterComponent, DestinyClass, DestinyGender } from 'bungie-api-ts/destiny2';
+import {
+  DestinyCharacterComponent,
+  DestinyClass,
+  DestinyGender,
+  DestinyProfileRecordsComponent,
+  DestinyRecordState,
+} from 'bungie-api-ts/destiny2';
 import vaultBackground from 'images/vault-background.svg';
 import vaultIcon from 'images/vault.svg';
 import { D2ManifestDefinitions } from '../../destiny2/d2-definitions';
 import { bungieNetPath } from '../../dim-ui/BungieImage';
-import { DimCharacterStat, DimStore } from '../store-types';
+import { DimCharacterStat, DimStore, DimTitle } from '../store-types';
 
 /**
  * A factory service for producing "stores" (characters or the vault).
@@ -21,7 +27,8 @@ const genderTypeToEnglish = {
 export function makeCharacter(
   defs: D2ManifestDefinitions,
   character: DestinyCharacterComponent,
-  mostRecentLastPlayed: Date
+  mostRecentLastPlayed: Date,
+  profileRecords: DestinyProfileRecordsComponent | undefined
 ): DimStore {
   const race = defs.Race[character.raceHash];
   const raceLocalizedName = race.displayProperties.name;
@@ -56,6 +63,9 @@ export function makeCharacter(
     genderName: genderTypeToEnglish[gender.genderType] ?? '',
     isVault: false,
     color: character.emblemColor,
+    titleInfo: character.titleRecordHash
+      ? getTitleInfo(character.titleRecordHash, defs, profileRecords, character.genderHash)
+      : undefined,
     items: [],
     hadErrors: false,
   };
@@ -116,4 +126,39 @@ export function getCharacterStatsData(
   });
 
   return ret;
+}
+
+function getTitleInfo(
+  titleRecordHash: number,
+  defs: D2ManifestDefinitions,
+  profileRecords: DestinyProfileRecordsComponent | undefined,
+  genderHash: number
+): DimTitle | undefined {
+  const titleRecordDef = defs?.Record.get(titleRecordHash);
+  if (!titleRecordDef) {
+    return undefined;
+  }
+  const title = titleRecordDef.titleInfo.titlesByGenderHash[genderHash];
+  if (!title) {
+    return undefined;
+  }
+
+  let gildedNum = 0;
+  let isGildedForCurrentSeason = false;
+
+  // Gilding information is stored per-profile, not per-character
+  if (titleRecordDef.titleInfo.gildingTrackingRecordHash) {
+    const gildedRecord =
+      profileRecords?.records[titleRecordDef.titleInfo.gildingTrackingRecordHash];
+
+    if (gildedRecord?.completedCount) {
+      gildedNum = gildedRecord.completedCount;
+    }
+
+    isGildedForCurrentSeason = Boolean(
+      gildedRecord && !(gildedRecord.state & DestinyRecordState.ObjectiveNotCompleted)
+    );
+  }
+
+  return { title, gildedNum, isGildedForCurrentSeason };
 }