+++ src/testing/test-utils.ts
@@ -9,7 +9,7 @@ import fs from 'fs/promises';
 import _ from 'lodash';
 import path from 'path';
 import { getManifest as d2GetManifest } from '../app/bungie-api/destiny2-api';
-import profile from './data/profile-2021-05-28.json';
+import profile from './data/profile-2022-06-27.json';
 
 /**
  * Get the current manifest as JSON. Downloads the manifest if not cached.